Code cleanup and fix all warnings for now

+++ src/a02/matrix.h
@@ -5,25 +5,22 @@
 #ifndef HPSC_MATRIX_H
 #define HPSC_MATRIX_H
 
+#include <cassert>
 #include <chrono>
 #include <cmath>
+#include <cstdint>
+#include <fstream>
+#include <functional>
+#include <iomanip>
 #include <iostream>
 #include <numeric>
 #include <sstream>
-#include <vector>
-#include <fstream>
-#include <functional>
-#include <src/common/utils.h>
 #include <thread>
-#include "Eigen/Eigen"
-#include <cassert>
-#include <cstdint>
-#include <iomanip>
-#include <iostream>
-#include <memory>
 #include <vector>
 
 #include "Eigen/Eigen"
+
+#include <src/common/utils.h>
 
 using EMatrix = Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic>;
 using namespace std;
@@ -39,22 +36,22 @@ class Matrix {
     }
 
     T& operator()(int n, int m) {
-      assert (n >= 0 && n < rows_);
-      assert (m >= 0 && m < cols_);
+      assert (n >= 0 && static_cast<uint32_t>(n) < rows_);
+      assert (m >= 0 && static_cast<uint32_t>(m) < cols_);
       return data_[n * cols_ + m];
     }
 
     T operator()(int n, int m) const {
-      assert (n >= 0 && n < rows_);
-      assert (m >= 0 && m < cols_);
+      assert (n >= 0 && static_cast<uint32_t>(n) < rows_);
+      assert (m >= 0 && static_cast<uint32_t>(m) < cols_);
       return data_[n * cols_ + m];
     }
 
     bool all_close(const Matrix<T> &other, bool fail_on_nan = true) const {
       assert (rows_ == other.rows_ && cols_ == other.cols_);
       T epsilon = 1e-4;
-      for(int i = 0; i < rows_; ++i) {
-        for(int j = 0; j < cols_; ++j) {
+      for(uint32_t i = 0; i < rows_; ++i) {
+        for(uint32_t j = 0; j < cols_; ++j) {
           if (std::fabs((*this)(i, j) - other(i, j) > epsilon)) {
             return false;
           }
@@ -85,15 +82,16 @@ class Matrix {
       return write_raw_rows(0, rows_, out, offset);
     }
 
-    int set_from(unique_ptr<T[]> &p, int offset = 0) {
+    /// Sets the values of this matrix from the given buffer. Useful for reading data received via MPI.
+    uint32_t set_from(unique_ptr<T[]> &p, uint32_t offset = 0) {
       return set_from(p.get(), offset);
     }
 
-    int set_from(T *raw, int offset = 0) {
-      int cur_offset = offset;
-      for(int i = 0; i < rows_; ++i) {
-        for(int j = 0; j < cols_; ++j) {
-          int idx = i * cols_ + j;
+    uint32_t set_from(T *raw, uint32_t offset = 0) {
+      uint32_t cur_offset = offset;
+      for(uint32_t i = 0; i < rows_; ++i) {
+        for(uint32_t j = 0; j < cols_; ++j) {
+          uint32_t idx = i * cols_ + j;
           data_[idx] = raw[cur_offset++];
         }
       }
@@ -128,8 +126,8 @@ class BandMatrix {
   }
 
   T& operator()(int row_id, int col_id) {
-    assert(row_id >= 0 && row_id < n_);
-    assert(col_id >= 0 && col_id < n_);
+    assert(row_id >= 0 && static_cast<uint32_t>(row_id) < n_);
+    assert(col_id >= 0 && static_cast<uint32_t>(col_id) < n_);
 
     if (abs(col_id - row_id) <= bandwidth_) {
       return data_.at(index_(row_id, col_id));
@@ -141,8 +139,8 @@ class BandMatrix {
 
   // TODO(andreib): Make this more consistent.
   T get(int row_id, int col_id) const {
-    assert(row_id >= 0 && row_id < n_);
-    assert(col_id >= 0 && col_id < n_);
+    assert(row_id >= 0 && static_cast<uint32_t>(row_id) < n_);
+    assert(col_id >= 0 && static_cast<uint32_t>(col_id) < n_);
 
     if (abs(col_id - row_id) <= bandwidth_) {
       return data_[index_(row_id, col_id)];
@@ -166,8 +164,8 @@ class BandMatrix {
   /// Returns a dense representation of the data in this matrix.
   Matrix<T> get_dense() const {
     std::vector<T> result_data;
-    for (int i = 0; i < n_; ++i) {
-      for (int j = 0; j < n_; ++j) {
+    for (uint32_t i = 0; i < n_; ++i) {
+      for (uint32_t j = 0; j < n_; ++j) {
         result_data.push_back(this->get(i, j));
       }
     }
@@ -225,14 +223,14 @@ template<typename T>
 Matrix<T> operator*(const Matrix<T> &left, const Matrix<T> &right) {
   assert(left.cols_ == right.rows_);
   std::vector<T> res_data;
-  for(int i = 0; i < left.rows_ * right.cols_; ++i) {
+  for(uint32_t i = 0; i < left.rows_ * right.cols_; ++i) {
     res_data.push_back(0.0);
   }
 
   Matrix<T> result(left.rows_, right.cols_, res_data);
-  for(int i = 0; i < left.rows_; ++i) {
-    for(int j = 0; j < right.cols_; ++j) {
-      for(int k = 0; k < left.cols_; ++k) {
+  for(uint32_t i = 0; i < left.rows_; ++i) {
+    for(uint32_t j = 0; j < right.cols_; ++j) {
+      for(uint32_t k = 0; k < left.cols_; ++k) {
         result(i, j) += left(i, k) * right(k, j);
       }
     }
@@ -271,7 +269,7 @@ Matrix<T> operator+(const Matrix<T> &left, const Matrix<T> &right) {
 /// Useful for printing a matrix as text to a stream. (Includes all zeros.)
 template<typename T>
 std::ostream& operator<<(std::ostream& out, const BandMatrix<T> &m) {
-  int n = m.get_n();
+  uint32_t n = m.get_n();
 
   for(uint32_t i = 0; i < n; ++i) {
     for(uint32_t j = 0; j < n; ++j) {
@@ -285,8 +283,8 @@ std::ostream& operator<<(std::ostream& out, const BandMatrix<T> &m) {
 
 template<typename T>
 std::ostream& operator<<(std::ostream& out, const Matrix<T> &m) {
-  for(int i = 0; i < m.rows_; ++i) {
-    for(int j = 0; j < m.cols_; ++j) {
+  for(uint32_t i = 0; i < m.rows_; ++i) {
+    for(uint32_t j = 0; j < m.cols_; ++j) {
       out << setw(6) << setprecision(4) <<  m(i, j);
       if (j != m.cols_ - 1 || m.cols_ == 1) {
         out << ", ";
@@ -304,8 +302,8 @@ std::ostream& operator<<(std::ostream& out, const Matrix<T> &m) {
 /// Prints a vector of printable things to a stream.
 template<typename T>
 std::ostream& operator<<(std::ostream& out, const std::vector<T>& data) {
-  int n = data.size();
-  for (int i = 0; i < n; ++i) {
+  uint64_t n = data.size();
+  for (uint32_t i = 0; i < n; ++i) {
     out << data[i];
     if (i < n - 1) {
       out << ", ";
@@ -314,7 +312,7 @@ std::ostream& operator<<(std::ostream& out, const std::vector<T>& data) {
   return out;
 }
 
-
+/// Generates a vector of the specified dimension full of zeros.
 vector<double> Zeros(int count);
 
 #endif //HPSC_MATRIX_H
